+++ app/frontend/src/features/auth/constants.tsx
@@ -1,19 +1,32 @@
 export const ABOUT_US = "About us";
 export const CHANGE_EMAIL = "Change Email";
+export const CHANGE_EMAIL_ERROR = "Error changing email: ";
+export const CHANGE_EMAIL_PROGRESS = "Email change in progress...";
+export const CHANGE_EMAIL_SUCCESS = "Your email has been changed successfully!";
 export const CHANGE_NAME_GENDER = "Change Name / Gender";
 export const CHANGE_PASSWORD = "Change Password";
+export const CHANGE_PASSWORD_ERROR = "Error changing password: ";
+export const CHANGE_PASSWORD_PROGRESS = "Password reset in progress...";
+export const CHANGE_PASSWORD_SUCCESS =
+  "Your password has been reset successfully!";
 export const CHECK_EMAIL =
   "Your email change has been received. Check your new email to complete the change.";
+export const CLICK_LOGIN = "Click here to log in";
 export const COLLABORATIVE = "Built by the community and always free";
 export const CONFIRM_PASSWORD = "Confirm password";
 export const CONTACT =
   "Admins will be happy to change your name and/or gender label. Please contact contact@couchers.org!";
 export const CURRENT_PASSWORD = "Current password";
+export const ENTER_EMAIL = "Enter your username/email";
 export const INTRODUCTION = "Connect with the world around you";
 export const LOGIN = "Log in";
 export const NEW_EMAIL = "New email";
 export const NEW_PASSWORD = "New password";
+export const OLD_PASSWORD = "Old password";
 export const PASSWORD_CHANGED =
-  "Your password change has been processed. Check your email for confirmation";
+  "Your password change has been processed. Check your email for confirmation.";
+export const RESET_PASSWORD = "Reset password";
+export const RESET_PASSWORD_LINK =
+  "Check your email for a reset password link!";
 export const SIGN_UP = "Sign up";
 export const SUBMIT = "Submit";
