+++ app/frontend/src/features/auth/email/ConfirmChangeEmail.tsx
@@ -1,13 +1,18 @@
 import { Typography } from "@material-ui/core";
+import Alert from "components/Alert";
+import {
+  CHANGE_EMAIL_ERROR,
+  CHANGE_EMAIL_PROGRESS,
+  CHANGE_EMAIL_SUCCESS,
+  CLICK_LOGIN,
+} from "features/auth/constants";
 import { Empty } from "google-protobuf/google/protobuf/empty_pb";
 import { Error as GrpcError } from "grpc-web";
 import { useEffect } from "react";
 import { useMutation } from "react-query";
 import { Link, useParams } from "react-router-dom";
-
-import Alert from "../../../components/Alert";
-import { loginRoute } from "../../../routes";
-import { service } from "../../../service";
+import { loginRoute } from "routes";
+import { service } from "service/index";
 
 export default function ConfirmChangeEmail() {
   const { resetToken } = useParams<{ resetToken?: string }>();
@@ -28,19 +33,17 @@ export default function ConfirmChangeEmail() {
   }, [completeChangeEmail, resetToken]);
 
   return isLoading ? (
-    <Typography variant="body1">Email change in progress...</Typography>
+    <Typography variant="body1">{CHANGE_EMAIL_PROGRESS}</Typography>
   ) : isSuccess ? (
     <>
-      <Alert severity="success">
-        Your email has been changed successfully!
-      </Alert>
+      <Alert severity="success">{CHANGE_EMAIL_SUCCESS}</Alert>
       <Typography variant="body1" component={Link} to={loginRoute}>
-        Click here to login
+        {CLICK_LOGIN}
       </Typography>
     </>
   ) : (
     error && (
-      <Alert severity="error">{`Error changing email: ${error.message}`}</Alert>
+      <Alert severity="error">{`${CHANGE_EMAIL_ERROR}${error.message}`}</Alert>
     )
   );
 }
