+++ app/frontend/src/features/auth/password/CompleteResetPassword.tsx
@@ -1,15 +1,20 @@
 import { Typography } from "@material-ui/core";
+import Alert from "components/Alert";
+import {
+  CHANGE_PASSWORD_ERROR,
+  CHANGE_PASSWORD_PROGRESS,
+  CHANGE_PASSWORD_SUCCESS,
+  CLICK_LOGIN,
+} from "features/auth/constants";
 import { Empty } from "google-protobuf/google/protobuf/empty_pb";
 import { Error as GrpcError } from "grpc-web";
 import { useEffect } from "react";
 import { useMutation } from "react-query";
 import { Link, useParams } from "react-router-dom";
+import { loginRoute } from "routes";
+import { service } from "service/index";
 
-import Alert from "../../../components/Alert";
-import { loginRoute } from "../../../routes";
-import { service } from "../../../service";
-
-export default function CompleteResetPasswordPage() {
+export default function CompleteResetPassword() {
   const { resetToken } = useParams<{ resetToken?: string }>();
 
   const {
@@ -28,20 +33,18 @@ export default function CompleteResetPasswordPage() {
   }, [completePasswordReset, resetToken]);
 
   return isLoading ? (
-    <Typography variant="body1">Password reset in progress...</Typography>
+    <Typography variant="body1">{CHANGE_PASSWORD_PROGRESS}</Typography>
   ) : isSuccess ? (
     <>
-      <Alert severity="success">
-        Your password has been reset successfully!
-      </Alert>
+      <Alert severity="success">{CHANGE_PASSWORD_SUCCESS}</Alert>
       <Typography variant="body1" component={Link} to={loginRoute}>
-        Click here to login
+        {CLICK_LOGIN}
       </Typography>
     </>
   ) : (
     error && (
       <Alert severity="error">
-        {`Error resetting password: ${error.message}`}
+        {`${CHANGE_PASSWORD_ERROR}${error.message}`}
       </Alert>
     )
   );
